apps/scorer/pyproject.toml
[project]
name = "scorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "glob2>=0.7",
    "pandas>=2.2.3",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "psycopg2-binary==2.9.10",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "scipy>=1.15.3",
]
